Minor refactoring of Transaction components
